Move config to a separate file
