+++ packages/nvidia_nat_mcp/src/nat/plugins/mcp/client_impl.py
@@ -17,18 +17,16 @@ import asyncio
 import logging
 from datetime import datetime
 from datetime import timedelta
-from typing import Literal
-
-from pydantic import BaseModel
-from pydantic import Field
-from pydantic import HttpUrl
-from pydantic import model_validator
+from typing import TYPE_CHECKING
 
 from nat.builder.builder import Builder
 from nat.builder.function import FunctionGroup
 from nat.cli.register_workflow import register_function_group
-from nat.data_models.component_ref import AuthenticationRef
-from nat.data_models.function import FunctionGroupBaseConfig
+from nat.plugins.mcp.client_config import MCPClientConfig
+from nat.plugins.mcp.client_config import MCPToolOverrideConfig
+
+if TYPE_CHECKING:  # Only for type checking; avoids runtime import/cycles
+    from nat.plugins.mcp.client_base import MCPBaseClient
 
 logger = logging.getLogger(__name__)
 
@@ -47,7 +45,7 @@ class MCPFunctionGroup(FunctionGroup):
         self._mcp_client_transport: str | None = None
 
         # Session management attributes for per-session isolation
-        self._session_clients: dict[str, object] = {}  # Will be MCPBaseClient instances
+        self._session_clients: dict[str, MCPBaseClient] = {}
         self._session_last_activity: dict[str, datetime] = {}
         self._cleanup_lock = asyncio.Lock()
 
@@ -126,11 +124,19 @@ class MCPFunctionGroup(FunctionGroup):
                 except Exception as e:
                     logger.warning("Error cleaning up session client %s: %s", session_id, e)
 
-    async def _get_session_client(self, session_id: str | None = None) -> object:
+    async def _get_session_client(self, session_id: str | None = None) -> MCPBaseClient:
         """Get the appropriate MCP client for the session."""
         if session_id is None:
             # Use default client for users without session_id
             return self.mcp_client
+
+        # If the session_id equals the configured default_user_id and default-user
+        # tool calls are allowed, use the base client instead of creating a per-session client
+        if self._shared_auth_provider and \
+                self._shared_auth_provider.config.allow_default_user_id_for_tool_calls:
+            default_uid = self._shared_auth_provider.config.default_user_id
+            if default_uid and session_id == default_uid:
+                return self.mcp_client
 
         if session_id not in self._session_clients:
             # Create session client lazily
@@ -146,7 +152,7 @@ class MCPFunctionGroup(FunctionGroup):
 
         return self._session_clients[session_id]
 
-    async def _create_session_client(self, session_id: str) -> object:
+    async def _create_session_client(self, session_id: str) -> MCPBaseClient:
         """Create a new MCP client instance for the session."""
         from nat.plugins.mcp.client_base import MCPSSEClient
         from nat.plugins.mcp.client_base import MCPStdioClient
@@ -218,106 +224,6 @@ class MCPFunctionGroup(FunctionGroup):
         return session_auth_provider
 
 
-class MCPToolOverrideConfig(BaseModel):
-    """
-    Configuration for overriding tool properties when exposing from MCP server.
-    """
-    alias: str | None = Field(default=None, description="Override the tool name (function name in the workflow)")
-    description: str | None = Field(default=None, description="Override the tool description")
-
-
-class MCPServerConfig(BaseModel):
-    """
-    Server connection details for MCP client.
-    Supports stdio, sse, and streamable-http transports.
-    streamable-http is the recommended default for HTTP-based connections.
-    """
-    transport: Literal["stdio", "sse", "streamable-http"] = Field(
-        ..., description="Transport type to connect to the MCP server (stdio, sse, or streamable-http)")
-    url: HttpUrl | None = Field(default=None,
-                                description="URL of the MCP server (for sse or streamable-http transport)")
-    command: str | None = Field(default=None,
-                                description="Command to run for stdio transport (e.g. 'python' or 'docker')")
-    args: list[str] | None = Field(default=None, description="Arguments for the stdio command")
-    env: dict[str, str] | None = Field(default=None, description="Environment variables for the stdio process")
-
-    # Authentication configuration
-    auth_provider: str | AuthenticationRef | None = Field(default=None,
-                                                          description="Reference to authentication provider")
-
-    @model_validator(mode="after")
-    def validate_model(self):
-        """Validate that stdio and SSE/Streamable HTTP properties are mutually exclusive."""
-        if self.transport == "stdio":
-            if self.url is not None:
-                raise ValueError("url should not be set when using stdio transport")
-            if not self.command:
-                raise ValueError("command is required when using stdio transport")
-            # Auth is not supported for stdio transport
-            if self.auth_provider is not None:
-                raise ValueError("Authentication is not supported for stdio transport")
-        elif self.transport == "sse":
-            if self.command is not None or self.args is not None or self.env is not None:
-                raise ValueError("command, args, and env should not be set when using sse transport")
-            if not self.url:
-                raise ValueError("url is required when using sse transport")
-            # Auth is not supported for SSE transport
-            if self.auth_provider is not None:
-                raise ValueError("Authentication is not supported for SSE transport.")
-        elif self.transport == "streamable-http":
-            if self.command is not None or self.args is not None or self.env is not None:
-                raise ValueError("command, args, and env should not be set when using streamable-http transport")
-            if not self.url:
-                raise ValueError("url is required when using streamable-http transport")
-
-        return self
-
-
-class MCPClientConfig(FunctionGroupBaseConfig, name="mcp_client"):
-    """
-    Configuration for connecting to an MCP server as a client and exposing selected tools.
-    """
-    server: MCPServerConfig = Field(..., description="Server connection details (transport, url/command, etc.)")
-    tool_call_timeout: timedelta = Field(
-        default=timedelta(seconds=60),
-        description="Timeout (in seconds) for the MCP tool call. Defaults to 60 seconds.")
-    auth_flow_timeout: timedelta = Field(
-        default=timedelta(seconds=300),
-        description="Timeout (in seconds) for the MCP auth flow. When the tool call requires interactive \
-        authentication, this timeout is used. Defaults to 300 seconds.")
-    reconnect_enabled: bool = Field(
-        default=True,
-        description="Whether to enable reconnecting to the MCP server if the connection is lost. \
-        Defaults to True.")
-    reconnect_max_attempts: int = Field(default=2,
-                                        ge=0,
-                                        description="Maximum number of reconnect attempts. Defaults to 2.")
-    reconnect_initial_backoff: float = Field(
-        default=0.5, ge=0.0, description="Initial backoff time for reconnect attempts. Defaults to 0.5 seconds.")
-    reconnect_max_backoff: float = Field(
-        default=50.0, ge=0.0, description="Maximum backoff time for reconnect attempts. Defaults to 50 seconds.")
-    tool_overrides: dict[str, MCPToolOverrideConfig] | None = Field(
-        default=None,
-        description="""Optional tool name overrides and description changes.
-        Example:
-          tool_overrides:
-            calculator_add:
-              alias: "add_numbers"
-              description: "Add two numbers together"
-            calculator_multiply:
-              description: "Multiply two numbers"  # alias defaults to original name
-        """)
-    session_aware_tools: bool = Field(default=True,
-                                      description="Session-aware tools are created if True. Defaults to True.")
-
-    @model_validator(mode="after")
-    def _validate_reconnect_backoff(self) -> "MCPClientConfig":
-        """Validate reconnect backoff values."""
-        if self.reconnect_max_backoff < self.reconnect_initial_backoff:
-            raise ValueError("reconnect_max_backoff must be greater than or equal to reconnect_initial_backoff")
-        return self
-
-
 def mcp_session_tool_function(tool, function_group: MCPFunctionGroup):
     """Create a session-aware NAT function for an MCP tool.
 
@@ -335,6 +241,10 @@ def mcp_session_tool_function(tool, function_group: MCPFunctionGroup):
         try:
             # Route to the appropriate session client
             session_id = function_group._get_session_id_from_context()
+
+            # If no session is available and default-user fallback is disabled, deny the call
+            if session_id is None:
+                return "User not authorized to call the tool"
             client = await function_group._get_session_client(session_id)
 
             # Resolve the tool from the routed client
@@ -448,7 +358,7 @@ async def mcp_client_function_group(config: MCPClientConfig, _builder: Builder):
                 tool_fn = mcp_session_tool_function(tool, group)
             else:
                 from nat.plugins.mcp.tool import mcp_tool_function
-                tool_fn = mcp_tool_function(tool)
+            tool_fn = mcp_tool_function(tool)
 
             # Normalize optional typing for linter/type-checker compatibility
             single_fn = tool_fn.single_fn
